requirements.txt
pyTelegramBotAPI==4.26.0
requests==2.32.5
python-dotenv==1.0.1
openai==1.66.3
google-generativeai
